* Add a resource manager suitable with SDL * Add standard C++ templates: Point, Rectangle & distances computations * Updated interface/* ,  game/* , team/* , map/* , object/* ,  weapon/ ( with only mine and dynamite for the momment)
This is a lot of changes but every change is inside preprocessor conditional (#ifndef CL)
I was obliged to change so many files at once because of internal dependencies... Porting each file implied changing nearly all the other ones.

Should compile and be "...playable at a very early stage" with SDL now.