Apply patch from Thibaud Reygner <thibouf AT gmail.com> to fix a uzi bug
"I have constated a little bug with the uzi : when you shoot with an exactly 90° angle , bullets are too speedy for this angle and make hole on the middle of a wall 
instead of the border.

An example on this screen : http://thibouf.free.fr/wormux/uzi%20bug.png

Th probleme is that for some angle, in each iteration, an incrementation of 1 in pos.X, pos.Y is too much incremented.

I have made a patch which correct by swithing between increment  pos.X or pos.Y in function of the angle. ( pos.Y for upwards and downwards,  pos.X else) .

I also made a little modification for checking if the bullet is outside of the map, because it did not check for the Y side and the bullet did not stop at the top 
of the map."