Fix bug #8529 - bazooka rocket rebounds on wall when in network
In fact, rocket does not rebound but are moved by the explosion (of itself) applied by the action handler (action has been sent over the network). 
The explosion arrives from the network and occurs just before the rocket collides the ground.

To fix the bug, a unique id is set for each WeaponProjectile and this unique_id is used in ApplyExplosion. If the unique_id of an object is the 
same as the unique_id associated to the explosion, nothing is done on the object.

The gameloop is responsible for attributing unique_id.