Partial revert of Kurosu commit about putenv/setenv
Setenv is POSIX too. Extract from man page:
CONFORMING TO
       4.3BSD, POSIX.1-2001.

Moreover, putenv take a char* argument not a const char*:
int putenv(char *string);

From the manpage:
The  libc4  and libc5 and glibc 2.1.2 versions conform to SUSv2: the pointer string given to putenv() is used.  In particular, this string becomes part of the environment; changing it later will change the environment. (Thus, it is an error is to call putenv() with an automatic  variable as the  argument,  then return from the calling function while string is still part of the environment.)  However, glibc 2.0-2.1.1 differs: a copy of the string is used.  On the one hand this causes a memory leak, and on the other hand it violates SUSv2. This has been fixed in glibc2.1.2.
    
The 4.4BSD version, like glibc 2.0, uses a copy.

SUSv2 removes the ‘const’ from the prototype, and so does glibc 2.1.3.