[MAJOR] Change the network random numbers generator management. Now each client compute its own random numbers but the seed is refreshed each turn by the turn master. This way, the random generators stay synchronized. This is related to bug #9861.
WARNING: The random generator seems unsynchronized in the initialization of the game, when mines and barrels are put. Thus, they are not in the same 
place. Anyway, now, when gnu and polecat rebounds, it is exactly at the same place, so it's ok for the generator, but there's something to fix during 
the initialization.