[BUGFIX] [NETWORK] bug #11231 -	Football Bomb is not network compatible
When a football bomb explodes, it create some other football bombs that were blasted by the explosion. In network game, it may have a little 
difference in the delay of the creation of the new football bombs and the explosion. Thus, the blasting effect on the bombs can be different from one 
client to another. 
To avoid to be blasted by it's own explosion in case of network game, we use an identifier associated to the physical object and the explosion. If an 
object has the same id than the explosion, the object is not blasted (cf bug #8529 - r3019). Here, we shared this id between all the football bombs, 
thus, they are not disturbed by explosions blast of themselves.