[MINOR] Remove some "stupid" comment
The comment was about the year information returned by time() (man 2). 
"time() returns the time since the Epoch (00:00:00 UTC, January 1, 1970), measured in seconds. "

BUT localtime() (man 3) fills a struct tm with the field tm_year equals to the "number of years since 1900."
For instance, it contains 98 for 1998, 108 for 2008. 

The comment was asking why we add 1900 and not 1970 to the tm.tm_year to get the year...