Use recv() instead of read() and send() instead of write() for input/output on socket. Add the MSG_NOWAIT flag to recv and the MSG_NOSIGNAL flag to send.
recv() is more adapted to socket than read and fails if the socket has been closed from the other side which is not the case of read if the SIGPIPE 
signal is blocked.

this is related to bug #12249 - Index server can be blocked in read()

I hope it fixes the bug!