bug #11545 - Change the behaviour of the floor on closed map
On closed map, characters were able to wake on the floor even if there is no more ground (if there is not water of course ;) )
This is a very strange behaviour, change it to have the same as in open map.
Remove PhysicalObj::FootsOnFloor, it is useless now.

(backport from trunk r5240)