[NETWORK][task #5662] DistantComputer contains a list of Players
When a wormux instance is client in a network game, it is connected to only one visible DistantComputer, that is the server. However, if there are
several clients connected to the server, from client A point of view, server is responsible of all other clients.
Class DistantComputer now has a list of Players, each Player corresponding to a network client (or server). Thus, each client has a real view of the
game, with the correct number of players.