[task #5662] Server handles now correctly several games :)
In particular, it detects when a game is started and can not accept new players. However, since it does not detect when a game is finished, there are
surely some problems when playing game in loop. Another problem is when we lost the game master and the elected one was waiting for game master
validation in the menu.