task #6459 : Make camera movement possible during weapon choice
it was already possible to move camera using ctrl+arrows or using mouse middle
button click but it was not possible to scroll putting the cursor near the edge
of the screen. This is now possible. The weapon menu is not anymore at the very
right of the screen if "scroll on border" option is enabled.

Patch copied from svn+ssh://svn.gna.org/svn/wormux/branches/wormux-0.8.4@5911