[BUGFIX] Handle network actions as soon as there are received when in network menu.
Actions are received from network by a thread, but they are executed by the
main thread that can be waiting for local SDL event (keyboard or mouse). Thus,
when waiting in the network menu, a network event (such as the remote selection
of a team) can be ignored for a long time and it can generate some bugs if it is
a network server...

Poll all SDL events as soon as we have received one from SDL_WaitEvent in the
menu. It avoids to redisplay several times the menu for handling several events
in a few times. It gives MUCH better performance when clicking.