[BUGFIX] Fix suppression of object from list of overlapped objects
This bug leads to a segfault when refreshing the list of objects. It was
possible to reproduce it using the m16 with lot of bullets per turn and shooting
again and again until the game crashes...

Thanks a lot to Ilari for his help in tracking and fixing the bug!