[BUGFIX] Fix access to options menu at startup
Since r6150, Wormux was crashing when trying to access Options menu without
having starting a game before.

To check that there is no game running before displaying the Options menu (Team
editor is not enable when a game is running) a Game object is created. The
creation of the Game object fails because the game mode is not yet loaded.

Modify the options menu to not create a Game object if there is none. In that
case, we are already sure that there is no game running.