[BUGFIX] Never try to create a Game object just to know if a game is running
Since r6150, Wormux was crashing when trying to access a network game without
having starting a local game before.

In several places, to check that there is no game running, a Game object is
created. The creation of the Game object fails because the game mode is not yet
loaded.

Modify class Game and callers to be sure that a Game object is never created
if there is none. In that case, we are already sure that there is no game
running.

This modification follows what has already been done in r6184.