[BUGFIX] Fix handling of key event during the game
When starting a game, key repeat is disabled from SDL and handled manually in
Wormux. Anyway, for menu, it is really useful to rely on SDL key repeat
handling and thus SDL key repeat is enabled when entering a menu.

Ignore KeyPressed event during the game on a key that is already recorded as
pressed. It means that key repeat has not been disabled...
Ignoring these invalid events avoids some strange gaming problem (such as
weapon strength bar not working, walking very slowly, ...)