[NETWORK][BUGFIX] Fix bug in unsynchronization of skin animation leading to assertion violation (and thus to a "crash")
The bug happens when a character dies at the very end of a turn, for some reason, on one computer, the character may have clothe Dead and movement Breathe while another may have clothe Dead and movement Black.

When forcing re-sync at the end of a turn, we ask the non-turn master computers to apply Clothe, Movement and Frame from the turn master using Character::SetClothe and Character::SetMovement. But, without option, Character::SetMovement does nothing if the character is dead.

As the character is dead, movement is not changed, and when requesting the frame of the movement, we get out of bounds, violating the assertion.

The bug is solved by forcing to apply clothe and movement, whatever state the character has.