[BUGFIX][Network] Exec actions AFTER refresh of objects has been done for each physic frame.
Previously, there was this following loop:
while (Game is running)
   RefreshTime
   RefreshInput // keyboard, mouse, ...
   RefreshActions // actions posted locally or from the network
   RefreshObjects // physical engine, game rules and their consequence

This is wrong, because an action generated by physical object at time
T is executed before the physical engine at time T, leading to
potential problem. For instance, Character::DoShoot may be cancelled
because SetStateHasPlayed Action has been received and is executed
before entering Character::DoShoot on the non turn master while on the
turn master, SetStateHasPlayed is the *consequence* of
Character::DoShoot

This patch modify the loop and make it:
While (Game is running)
    RefreshTime
    RefreshInput
    RefreshObjects
    RefreshActions

That partially fix bug #11766 (Some weapon are not draw).

BUT, it has some drawback: it delays the execution of actions and thus
it can be executed too late. Before it can already be executed too
late and sometimes a little bit in advance.

To make sure actions are executed at the right time, iFlo patchset is
a good track.