Don't continue the game if there could be outstanding actions.
The game menu does now pause the game in network games as well:
If the turn master pauses then the game will stop for the non turn masters as they don't get the permission to calculate further frames.
If a non turn mater pauses the game then the game continues for the other players. Once the non turn master resumes the game it will continue for him, but he will be behind in time.
If a player is behind in tame then the other players will wait for him when he becomes turn master.

Non game related actions get executed anyway which makes it possible to receive chat messages instantly even if you are behind in time.