Fix a bug which leads to different water heights in network games.
The position of water waves must not depend on how often the waves get drawn.
Especially not since an optimization does not draw the waves if they are not visible.

The position of the waves is game relevant as it determines how high the water is at specific x positions.
The height of the water influences the game:
For example determines the water height when a footbomb will fly slower due to water resistance and thus when the footbomb will explode.