[BUGFIX] Fallback to previous screen resolution if a failure happens when applying the new selected one
If the failure happens at the game initialization, we try to apply a default screen
configuration of 640x480 with hardware acceleration.
It allows user with a wormux config file not corresponding to their hardware (because
of computer change or previous selection that was not good) to start Wormux.

Imported from: svn+ssh://svn.gna.org/svn/wormux/branches/wormux-0.8.5@6183
Original author: gentildemon
Original date: Fri Jul 17 08:21:10 2009 +0000