Verify the random seed at the beginning of each frame.
Previously the verification has been been scheduled in SetState of the previous frame.

This patch causes regressions as it assumes that there are no "recations".
Which reactions I mean in this case actions which get scheduled as a result of the physics frame calculaton.

Follow up patches will remove those reactions.