Make character movements transferable over network.
Previous this patch the character got moved by the HandleKeyRefreshed_Move* methods.
Sending a character movement notification every frame would cause a lot of network traffic.
This patch let the turn master only send two actions for a simple movement:
One which starts the movement and one which stops it.