Use NewAction for jumps.
Movement orders get transferred using NewAction.
If a user press "jump" and "move left" simultaneously while the character looks right then it matters if the character turns or jumps first.
Thus it must not happen that the turn master schedules the "move left" and executes the "jump" without scheduling. In such a case the non tun master would execute it the other way round as he received the jump order later.