Don't build the body in the draw method.
If the body gets rebuild in the draw method then the hand position gets dependent of the draw rate.
With the hand position the behavior of grapple gets dependent of the draw rate.
Thus players will get different results in network games if the body gets build in the draw method.
That's why the characters gets now build at other locations in the code.