Fix refactor of  6698.
The variables a and b got used by both code parts. They were missing for the draw method, but two constants a and b exists who were then used instead. As a result the water was to high to fit in the pattern and it looked like as it were clipped.