Handle leavers for all players at the same time.
The DisconnectPlayer method gets called by a frameless action and thus at different times for different players.
That's why it's a bad idea to remove the team in this disconnect method.

It's save to remove a leaver's team when we are waiting for more actions from the leaver.
The server can't have more actions buffered for us, as it sent the disconnect notification after forwarding the last action to us.

If a non turn master leaves then this doesn't have any effect until the leaver becomes turn master.
I think this improves game experience: If you are turn master then you want to finish your turn and don't want your targets to be removed right before your nose.