Revert revision 7000 as it cause regessions.
Revision 7000 changes Refresh() of the character class to only execute if the character is visible.
Currently the Refresh() method has the following task:
  Calculate the next physics frame for the selected character.

Thus revision 7000 makes the camera determine what happens in the game.
For example will characters stop moving or jumping if they are out of view.
This is not only a normal bug but also causes network games to desynchronize as players have different cameras.

Refresh() should not be dependent of the characters visiblity. If some calculations are only necessary for drawing then they should happen in the Draw() method.

If you want to move stuff from Refresh() to Draw() then please test it carefully in network games. Expecially as for example the fact if a character is build or not has influence on the game.