Teach the new AI how to use the bazooka.
It doesn't use the mechanism of the old AI of determining the fire angle and the weapon strength.
Instead the algorithm can find an fire angle and strength for any flight duration given.
The AI has also now a method to detect what a rocket missile will most likely hit.
It will shoot now with the bazooka at any reachable enemy which can be hit after 1 or 3 seconds flight duration.
If no enemy can hit directly it will shoot at ground destroying slowly the ground which protects the enemy.