Don't move the character while using blowtorch.
The user has now to press the arrow keys to move the character.
The advantage is that it looks more natural how the character moves: The character doesn't pop forwards at each blow.
Additionally it's not necessary anymore to use some character methods and fields which should not be public.