Reduce the amount of walk methods in character.
There were to many methods related to character movement within the character class.
For example were the StartMovingLeft, StartWalking and BeginMovementRL methods which get all called if you started moving left.
This refactor introduce a class WalkIntention which remembers what the user wants to do.
The remaining StartWalking method gets now only called, if the character actually starts walking.
For example could it be that there is the intention to walk, but the character can't walk as it's in the air.