Don't use IsInUse() to determine if a weapon should be displayed or not.
This patch makes it more clear when a weapon is actually visible and when not.
Previously the visibility could be made dependent of the return value of IsInUse(). That IsInUse gets used to determine the visibility isn't obvious as the enum talks of "WHEN_ACTIVE".
IsInUse gets overriden quite often which makes it even less obvious when a weapon gets displayed and when not.

Last but not least is IsInUse a method which gets called for a lot of different purposes.
So another developer might change the behavior of IsInUse to fix a unrelated issue while changing when the weapon is visible.