Backport "Don't spam DNS requests when running in debug mode."
Previously a debug build of wormux would have tried to resolve the sockets ip very frequently in order to get the address for a log message.
This commit fixes the following bug:
bug #15043: Game waits for a player, continues, waits for a player and continues again in a long loop.