Make it possible to weight how weapon usages get rated.
The AI is too good with the bazooka. This mechanic makes it possible to let the bazooka calculate a better fitting rating, while still ensuring that the AI does not use the bazooka always.
There is now a small chance that the AI use the bazooka at the start of the game.