Make the AI focus on one character when ground is in the way.
Before this patch the AI was rating a strategy to shoot B higher then usual as the explosion was close to A.
Now only the distance to the target counts when a ground bonus gets calculated.

This also fixes the issue that the AI was always preferring the grenade over the gun.
The bugs origin was the ground removal bonus each grenade explosion had.