Fix a desynchronization problem with Anvil.
This patch fixes the following bug:
bug #15143: Anvil destroys sometimes synchronization when it hits.

The variable explode_with_timeout was true which made the projectile use begin_time which wasn't initialized.
The uninitialized values of begin time can cause WeaponProjectile::IsImmobile() to return true for one player and false for the other.