Ignore actions about unknown players.
It can happen that a player receives an "add/remove team X for player Y" action even if Y has already disconnected.

This can happen for example if player A requests a team and disconnects.
The game master receives the request for the team and sends an "add team for A" action.
The server however has already told a player B that A has disconnected when it forwards the "add team for A" action to B.
In that case B has removed A already from it's player list when it receives "add team for A" action.