Move the player related state field from DistantComputer to Player.
Previously the state field of the DistantComputer contained the state of one of it's players.
Which player depended on a lot of factors and the state of the other players got not saved at all.

This patch fixes the following bug:
bug #15159: Assertions fail if the game master tries to start the game on a headless server when all but one player are ready.