bug #15271 - Fix shoot key handling when user exceeds the weapon max strength
The original bug with the automatic bazooka was due to an unsigned int
overflow (unit-- with unit == 0). It was leading to an assertion failure
in debug mode.

The real problem is that we generate a Action_Weapon_StopShooting when
user releases the shoot key even if shoot has been already done because
user exceeds the weapon max power, triggering the shoot.