bugfix - fix rope size check when attaching
Max rope length was badly checked when trying to attach because of an
integer overflow (uint lenght (== 0) - skip (==4)).
Thus, it was only checked while trying to increase the rope size. The
result is that you can attach from very far, then decrease the rope size
and then be unable to increase it to its previous length.

Moreover, the attaching was instantaneously during the shoot.

This patch fixes the rope size check and remove the code that delays the
attaching until the rope is long enough. The code was a bit crappy, was
not functionnal (maybe for a very long time), and it seems to lack to
nobody.