Make poison deadly for characters when they have to less health.
Previously there was no reason for a player to use it:
He or she needs one turn to poison the enemy and another plus a lot of waiting time to kill the enemy.
No smart player used to do that as she or he could archive the same with a simple gun much faster and from distance.