Ensure that the systems stdint.h file gets used whenever possible.
The fixedpoint library itself tries to include the systems stdint.h file for most systems.
It takes however the stdint.h file of the fixedpoint directory as the fixedpoint library is in our include path.
The rename ensures that the systems stdint.h file gets used.

The usage of the systems stdint.h file solves compile errors on 64 bit platforms and on mac.